Bump to Go 1.23 and latest linter (#68)
* Bump to Go 1.23 and latest linter
* covet: Fix lint failures on Go 1.23
* gopages: Fix lint failures on Go 1.23
* datasize: Fix lint failures on Go 1.23
* dns: Fix lint failures on Go 1.23
* goop: Fix lint failures on Go 1.23
* pipe: Fix lint failures on Go 1.23
* dns: Use latest x/net

+++ covet/cmd/covet/run.go
@@ -230,7 +230,7 @@ func findFirstUncoveredLines(lines []covet.Line, startIndex int) (uncovered span
 	for _, l := range lines[nextLineIndex:] {
 		nextLineIndex++
 		if !l.Covered {
-			n := int64(l.LineNumber)
+			n := l.LineNumber
 			uncovered = span.Span{
 				Start: n,
 				End:   n + 1,
@@ -241,7 +241,7 @@ func findFirstUncoveredLines(lines []covet.Line, startIndex int) (uncovered span
 	}
 	// find next line number jump or covered line
 	for _, l := range lines[nextLineIndex:] {
-		if l.Covered || int64(l.LineNumber) != uncovered.End {
+		if l.Covered || l.LineNumber != uncovered.End {
 			break
 		}
 		nextLineIndex++
